requirements.txt
astroid==2.0.4
autopep8==1.4
Django==2.1.2
django-ckeditor==5.6.1
django-extensions==2.1.3
django-js-asset==1.1.0
django-taggit==0.23.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pylint==2.1.1
pytz==2018.5
six==1.11.0
wrapt==1.10.11
dj-database-url==0.5.0
django-bootstrap3==10.0.1
django-environ==0.4.5
django-filter==2.0.0
gunicorn==19.9.0
psycopg2==2.7.5
raven==6.9.0
whitenoise==4.0
pillow
mysqlclient
